akida: Force 32bit PCIe accesses on Raspberry Pi CM4 (BCM2711)
64bit PCIe accesses are not supported on Raspberry Pi CM4
https://lore.kernel.org/linux-arm-kernel/20210222175549.GO1463@shell.armlinux.org.uk/T/

The Raspberry Pi CM4 uses a BCM2711.
This platform is detected using CONFIG_ARCH_BCM2835 even if the
SOC used is a BCM2711. From the Kconfig help section related to
ARCH_BCM2835:
  This enables support for the Broadcom BCM2711 and BCM283x SoCs.
  This SoC is used in the Raspberry Pi and Roku 2 devices.

So, force 32bit accesses if CONFIG_ARCH_BCM2835 is set.

The limitation with this method is that all kernel compiled to
suppport this soc will lead to 32bit PCIe accesses. This can
happen with kernel supporting multiple SOCs.
But a kernel will never support multiple architectures (ie ARM
and x86_64 in the same kernel binary).

As a local Akida workaround to support the Raspberry Pi CM4, this
is acceptable.

Signed-off-by: Herve Codina <herve.codina@bootlin.com>