test: Enable optimizations
There is no reason to compile without optimizations.

For these test programs, a compilation with optimisations ensure
that everything is coded correctly in each program.
It will be better to detect issues related to some optimization in test
program rather than detecting these issues in the final application.

Switch to '-O3'

Signed-off-by: Herve Codina <herve.codina@bootlin.com>